app/requirements.txt
python-dateutil
fastapi
uvicorn
unittest
pytest
pytest-cov
coveralls